requirements.txt
opencv-python
mediapipe
scikit-learn
pandas
numpy
matplotlib
pillow
